libraries/dagster-evidence/pyproject.toml
[project]
name = "dagster-evidence"
description = "Dagster integration with evidence.dev"
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "dagster>=1.12.10",
    "gitpython>=3.1.0",
    "pyyaml>=6.0",
]
dynamic = ["version"]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-mock",
]

[tool.uv]
dev-dependencies = [
    "ruff",
    "pytest",
    "pytest-mock",
    "pyright>=1.1.386",
]

[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["dagster_evidence", "dagster_evidence.lib", "dagster_evidence.components"]

[tool.setuptools.dynamic]
version = {attr = "dagster_evidence.__version__"}

[project.entry-points]
"dagster_dg.library" = { dagster_evidence = "dagster_evidence"}
"dagster_dg.plugin" = { dagster_evidence = "dagster_evidence"}
